add capabilityList
